ProjetoCamila/requirements.txt
fastapi
uvicorn
paho-mqtt
mysql-connector-python
python-dotenv
pydantic
